requirements.txt
requests
newspaper3k
beautifulsoup4
Flask